feat(auth): refactor user registration to assign default role
Refactor the registration process in the AuthController to assign a default role of 'user' to newly registered users. This is done by removing the explicit assignment of the role_id property and instead retrieving the role entity with the name 'user' from the roles repository. The retrieved role's ID is then used to set the role_id property of the user entity before saving it to the database.

Related to #123